Run release on release creation (#48)
Summary:
Pushing tags against existing commits doesn't seem to trigger the action, so let's list releases here instead.

Docs for this in D26607139.

Pull Request resolved: https://github.com/facebookincubator/fbjni/pull/48

Test Plan:
Ran it on my fork. It failed but because there were unclosed previous updates, which is in fact expected (and quite useful, too, because it means we don't have a stray release).

https://github.com/passy/fbjni/runs/1961026784?check_suite_focus=true

Reviewed By: nikoant

Differential Revision: D26583351

Pulled By: passy

fbshipit-source-id: 1eee0b8c5b294a778da95c7017a9fbe4f5232354