Remove the package exclusions
Summary:
It's all a little confusing but downstream consumers of Flipper
that don't use CMake themselves otherwise don't have access to these
libraries and will have them missing from their builds.

I'll create an 0.2.0 release for this so 0.1.0 consumers don't
end up with multiple binaries.

Test Plan:
Tested locally with Flipper and a stand-alone example and it seems to
work.