0.5.1 Correctly publish the -java-only artifact and the -headers artifact
Summary:
Fixes #85

I've realized that 0.5.0 was published successfully, but the `fbjni-java-only` artifact and the `headers`
artifacts are missing. This is because the publishing plugin was not applied to the host build, and
changes on the maven publish plugin required a bit more changes on the build files.

I've fixed them all and verified on maven local that everything publishes correctly.

I'm also bumping the version to 0.5.1 to publish a new version.

Reviewed By: cipolleschi

Differential Revision: D47718649

fbshipit-source-id: bcc95fb6011867d61bbeac26c9f61b4149bf0e49