Use a separate executor/thread for ClientSync
Summary: Sometimes ClientSync never gets created because the relevant task, which is scheduled on an executor, is never called. We do not know the root cause, but we (Kay, Michael, JP, and I) attempt to minimize the severity of the issue by creating a separate executor/thread for ClientSync on every new TW cast session.

Differential Revision: D51184111

fbshipit-source-id: 7fa7265f928671d95c1791eb111707584fb16efe