Remove libstdc++ support
Summary:
Android only uses libc++ now. The libc++ implementation should also work
for libstdc++ though (for other platforms if desired). The libc++
implementation doesn't rely on any headers not available on Windows so
we can remove the `_WIN32` conditionals (though it still won't work on
Windows because of ABI library differences, but that's not different
from before).

Reviewed By: thevinster

Differential Revision: D58622935

fbshipit-source-id: aa55809d570101c1be97184feb8255aa61bb66eb