Plumb binder from proxy
Summary:
This creates the BnSharedActivityApi instance in the proxy impl and plumbs it to the CallProviderService. This will then allow clients to get the API when they bind to the CallProviderService.

The ISharedActivityApi is just a skeleton for now and will be filled in over time.

Differential Revision:
D63823238

Privacy Context Container: L1272839

fbshipit-source-id: 1c9b6d40c91357cd3c18053472b20a0db5492866