Discovery C++ Interop
Summary: This creates a temporary wrapper around the CallDiscoveryManager library for the shared activities prototype. It adapts the flow-based API to a callback api for JNI/C++. Longer term we should land a generic adapter for KT flow to c++ unifex over JNI, but that is outside of the scope of this prototype effort.

Differential Revision:
D65771815

Privacy Context Container: L1272839

fbshipit-source-id: 1519e41c476d6e63fe5a2ccf745709c3c29368df