chore: update deps (#103)
Summary:
## Motivation

Bumping gradle, AGP and compile targets to match RN's core (Hence why i didn't go for gradle 8.12 or a higher AGP). Mostly just keeping things in sync, and i wanted to know if there was a specific reason this library was still targeting Java 8.

1. Update gradle to 8.11.1
2. Update `com.android.library` to 8.7.2 (why are we using this instead of specifically using AGP? I see https://github.com/facebookincubator/fbjni/commit/7621c81d3b3f6b7abf541a973b3f04f69e5fbe9d, but my android knowledge is not good enough to know why this was changed to the library itself.)
3. Update the Java target to 17 to see if the CI would run. Looking at the action, it sets up with 17 anyway

Pull Request resolved: https://github.com/facebookincubator/fbjni/pull/103

Test Plan:
CI. When running the tests locally, it complains about it targeting x86, but finding arm64 instead. Guessing it is either only made to run on the CI, or i missed an important step in the contribution guide
<img width="1155" alt="Screenshot 2025-01-03 at 03 42 35" src="https://github.com/user-attachments/assets/777eda61-b892-40cd-90d8-41b476a6c5c4" />

Reviewed By: passy

Differential Revision: D67793253

Pulled By: cortinico

fbshipit-source-id: 485c84045dbfd9675e098c49099d73c5bd62b2a9