Remove unused-variable in ../fbandroid/libraries/fbjni/cxx/fbjni/detail/Environment.cpp +1
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D68509706

fbshipit-source-id: 2e1390ea449a6f85b612f27a99425f4ca9a2ca80